Complete the proposal_layer

+++ Faster-RCNN-Practice/libs/cpu_nms.pyx
@@ -16,20 +16,20 @@ Created on Fri Jan 20 15:42:14 2017
 import numpy as np
 cimport numpy as np
 
-cdef inline np.float32_t max(np.float32_t a, np.float32_t b):
+cdef inline np.float64_t max(np.float64_t a, np.float64_t b):
     return a if a >= b else b
 
-cdef inline np.float32_t min(np.float32_t a, np.float32_t b):
+cdef inline np.float64_t min(np.float64_t a, np.float64_t b):
     return a if a <= b else b
 
-def cpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
-    cdef np.ndarray[np.float32_t, ndim=1] x1 = dets[:, 0]
-    cdef np.ndarray[np.float32_t, ndim=1] y1 = dets[:, 1]
-    cdef np.ndarray[np.float32_t, ndim=1] x2 = dets[:, 2]
-    cdef np.ndarray[np.float32_t, ndim=1] y2 = dets[:, 3]
-    cdef np.ndarray[np.float32_t, ndim=1] scores = dets[:, 4]
+def cpu_nms(np.ndarray[np.float64_t, ndim=2] dets, np.float thresh):
+    cdef np.ndarray[np.float64_t, ndim=1] x1 = dets[:, 0]
+    cdef np.ndarray[np.float64_t, ndim=1] y1 = dets[:, 1]
+    cdef np.ndarray[np.float64_t, ndim=1] x2 = dets[:, 2]
+    cdef np.ndarray[np.float64_t, ndim=1] y2 = dets[:, 3]
+    cdef np.ndarray[np.float64_t, ndim=1] scores = dets[:, 4]
 
-    cdef np.ndarray[np.float32_t, ndim=1] areas = (x2 - x1 + 1) * (y2 - y1 + 1)
+    cdef np.ndarray[np.float64_t, ndim=1] areas = (x2 - x1 + 1) * (y2 - y1 + 1)
     cdef np.ndarray[np.int_t, ndim=1] order = scores.argsort()[::-1]
 
     cdef int ndets = dets.shape[0]
@@ -41,11 +41,11 @@ def cpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
     # sorted indices
     cdef int i, j
     # temp variables for box i's (the box currently under consideration)
-    cdef np.float32_t ix1, iy1, ix2, iy2, iarea
+    cdef np.float64_t ix1, iy1, ix2, iy2, iarea
     # variables for computing overlap with box j (lower scoring box)
-    cdef np.float32_t xx1, yy1, xx2, yy2
-    cdef np.float32_t w, h
-    cdef np.float32_t inter, ovr
+    cdef np.float64_t xx1, yy1, xx2, yy2
+    cdef np.float64_t w, h
+    cdef np.float64_t inter, ovr
 
     keep = []
     for _i in range(ndets):
